+++ EventPlanner/src/app/offering/model/pricelist-model.ts
@@ -1,0 +1,8 @@
+export interface PricelistItem {
+    id: number;
+    offeringId: number;
+    name: string;
+    price: number;
+    discount?: number;
+    priceWithDiscount: number;
+  }
