+++ EventPlanner/src/app/offering/pricelist/pricelist.component.spec.ts
